TASK_2/requirements.txt
google-generativeai
pytesseract
Pillow
PyPDF2
pdf2image
